test: cover notification templates and telegram tasks, split oversized test files
Tech-debt sweep findings 24, 25, 26. Tests only; no production code changed.

Finding 24: five notification template modules and three telegram modules
executed only their import/class/def lines; every body line was unreached
because test_channels.py stubs get_template/get_notification_keyboard and
src/conftest.py autouse-mocks telegram.tasks.send_message_task. Most
importantly, telegram/tasks.py:44-51, the allowlist that guards an
importlib.import_module + getattr on Celery-supplied strings, had zero
executed statements.

- notifications/tests/test_template_registry_rendering.py: every
  NotificationType resolves through the real registry and renders a
  non-empty, fully-substituted title and subject; potluck action x
  organizer x actor branches and the waitlist ngettext plural are pinned.
- telegram/tests/test_tasks.py: the allowlist deny path first (arbitrary
  modules and both half-matches of the allowed pair never reach
  import_module), then the allow path, delivery-status callbacks, the
  blocked/deactivated/rate-limit/re-raise branches, and a bot smoke test.
- telegram/tests/test_notification_keyboards.py: the real keyboard
  builders, with callback-data round-trips for the organizer actions.

Branch coverage on the eight modules goes from body-zero to 100% (both
template files and telegram/tasks.py), 98% (notification_keyboards.py,
one unreachable line) and 58% (bot.py, run_bot's polling loop excluded).

Finding 25: hoist the byte-identical org_owner/org/member_user/event
fixtures in test_ticket_service.py to module scope (24 duplicate copies
removed, -183 lines). The deliberately different event override in
TestCheckInTicketAtTheDoor is kept and still shadows the module fixture.

Finding 26: split the five test files sitting just under the 1000-line cap
on class-cluster boundaries:
- test_subscription_revival/ (conftest, service, endpoints)
- test_attendee_invoice/ (_helpers, generation, lifecycle);
  test_attendee_credit_note_service.py imports the helpers from _helpers
- test_service/test_organization_service/ (conftest for the shared token
  fixtures; membership, lifecycle and token clusters)
- test_signals_follow/ (its single 900-line class splits at its natural
  seam into recipient-selection and trigger/payload rules, no method
  reordered)
- test_organization_admin/test_subscriptions.py -> two siblings in the
  same directory (it depends on the parent conftest chain), with the
  fixture block duplicated per that directory's established convention

Test counts are identical before and after for every split file.
