--- src/events/tests/test_attendee_invoice_service.py
@@ -1,992 +1,0 @@
-"""Tests for the attendee invoice service: validation, generation, editing, credit notes.
-
-Tests cover:
-- validate_invoicing_prerequisites() -- all 4 prerequisite checks
-- set_invoicing_mode() -- NONE always allowed, HYBRID/AUTO require prerequisites
-- generate_attendee_invoice() -- HYBRID/AUTO modes, idempotency, edge cases
-- update_draft_invoice() -- editable fields, disallowed fields, PDF invalidation
-- issue_draft_invoice() -- DRAFT -> ISSUED transition
-- delete_draft_invoice() -- draft deletion and issued rejection
-
-Credit-note tests (generate_attendee_credit_note) live in
-test_attendee_credit_note_service.py.
-"""
-
-import typing as t
-from decimal import Decimal
-from unittest.mock import patch
-
-import pytest
-from django.utils import timezone
-from ninja.errors import HttpError
-
-from accounts.models import RevelUser
-from events.models import Event, Organization, Ticket, TicketTier
-from events.models.attendee_invoice import AttendeeInvoice
-from events.models.ticket import Payment
-from events.service.attendee_invoice_draft_service import (
-    DERIVED_TOTAL_FIELDS,
-    delete_draft_invoice,
-    issue_draft_invoice,
-    update_draft_invoice,
-)
-from events.service.attendee_invoice_service import (
-    _is_export,
-    generate_attendee_invoice,
-    set_invoicing_mode,
-    validate_invoicing_prerequisites,
-)
-
-pytestmark = pytest.mark.django_db
-
-MOCK_RENDER_PDF = "events.service.attendee_invoice_service.render_pdf"
-MOCK_SEND_EMAIL = "common.tasks.send_email"
-
-
-# ---------------------------------------------------------------------------
-# Shared helpers
-# ---------------------------------------------------------------------------
-
-
-def _make_org_invoicing_ready(org: Organization) -> Organization:
-    """Configure an org with all prerequisites for invoicing."""
-    org.vat_country_code = "IT"
-    org.vat_id = "IT12345678901"
-    org.vat_id_validated = True
-    org.vat_rate = Decimal("22.00")
-    org.billing_name = "ACME SRL"
-    org.billing_address = "Via Roma 1, 00100 Roma"
-    org.billing_email = "billing@acme.it"
-    org.contact_email = "info@acme.it"
-    org.save()
-    return org
-
-
-def _create_payment(
-    *,
-    user: RevelUser,
-    event: Event,
-    tier: TicketTier,
-    session_id: str = "cs_test_123",
-    amount: Decimal = Decimal("100.00"),
-    net_amount: Decimal | None = Decimal("81.97"),
-    vat_amount: Decimal | None = Decimal("18.03"),
-    vat_rate: Decimal | None = Decimal("22.00"),
-    buyer_billing_snapshot: dict[str, t.Any] | None = None,
-    guest_name: str = "Test Guest",
-) -> Payment:
-    """Create a ticket and payment for testing."""
-    ticket = Ticket.objects.create(event=event, user=user, tier=tier, guest_name=guest_name)
-    return Payment.objects.create(
-        ticket=ticket,
-        user=user,
-        stripe_session_id=session_id,
-        status=Payment.PaymentStatus.SUCCEEDED,
-        amount=amount,
-        net_amount=net_amount,
-        vat_amount=vat_amount,
-        vat_rate=vat_rate,
-        platform_fee=Decimal("5.00"),
-        currency="EUR",
-        buyer_billing_snapshot=buyer_billing_snapshot,
-    )
-
-
-def _default_billing_snapshot(reverse_charge: bool = False) -> dict[str, t.Any]:
-    return {
-        "billing_name": "Buyer GmbH",
-        "vat_id": "DE123456789",
-        "vat_country_code": "DE",
-        "vat_id_validated": True,
-        "billing_address": "Berliner Str. 1, 10115 Berlin",
-        "billing_email": "buyer@example.de",
-        "reverse_charge": reverse_charge,
-    }
-
-
-_INVOICE_COUNTER = 0
-
-
-def _create_draft(org: Organization, event: Event, user: RevelUser) -> AttendeeInvoice:
-    """Create a minimal draft invoice for testing."""
-    global _INVOICE_COUNTER  # noqa: PLW0603
-    _INVOICE_COUNTER += 1
-    return AttendeeInvoice.objects.create(
-        organization=org,
-        event=event,
-        user=user,
-        stripe_session_id=f"cs_draft_{_INVOICE_COUNTER}",
-        invoice_number=f"TEST-{_INVOICE_COUNTER:06d}",
-        status=AttendeeInvoice.InvoiceStatus.DRAFT,
-        total_gross=Decimal("100.00"),
-        total_net=Decimal("81.97"),
-        total_vat=Decimal("18.03"),
-        vat_rate=Decimal("22.00"),
-        currency="EUR",
-        # Reconciles with the totals above: since #911 an invoice whose header
-        # disagrees with its lines cannot be issued, and a fixture that drifts
-        # is not a shape production can produce.
-        line_items=[
-            {
-                "description": "Event — Tier — Guest",
-                "unit_price_gross": "100.00",
-                "discount_amount": "0.00",
-                "net_amount": "81.97",
-                "vat_amount": "18.03",
-                "vat_rate": "22.00",
-            }
-        ],
-        seller_name="ACME SRL",
-        seller_email="billing@acme.it",
-        buyer_name="Original Buyer",
-        buyer_email="buyer@example.com",
-    )
-
-
-def _create_issued(org: Organization, event: Event, user: RevelUser) -> AttendeeInvoice:
-    """Create an issued invoice for testing."""
-    inv = _create_draft(org, event, user)
-    inv.status = AttendeeInvoice.InvoiceStatus.ISSUED
-    inv.issued_at = timezone.now()
-    inv.save()
-    return inv
-
-
-# ---------------------------------------------------------------------------
-# validate_invoicing_prerequisites
-# ---------------------------------------------------------------------------
-
-
-class TestValidateInvoicingPrerequisites:
-    """Test all 4 prerequisite checks for enabling invoicing."""
-
-    def test_passes_when_all_prerequisites_met(self, organization: Organization) -> None:
-        """No error raised when org has all required fields."""
-        validate_invoicing_prerequisites(_make_org_invoicing_ready(organization))
-
-    def test_fails_when_vat_country_code_missing(self, organization: Organization) -> None:
-        """Missing EU VAT country code should fail."""
-        org = _make_org_invoicing_ready(organization)
-        org.vat_country_code = ""
-        org.save()
-        with pytest.raises(HttpError) as exc_info:
-            validate_invoicing_prerequisites(org)
-        assert exc_info.value.status_code == 422
-        assert "EU-based" in str(exc_info.value)
-
-    def test_fails_when_vat_country_code_non_eu(self, organization: Organization) -> None:
-        """Non-EU VAT country code should fail."""
-        org = _make_org_invoicing_ready(organization)
-        org.vat_country_code = "US"
-        org.save()
-        with pytest.raises(HttpError) as exc_info:
-            validate_invoicing_prerequisites(org)
-        assert exc_info.value.status_code == 422
-
-    def test_fails_when_vat_id_not_validated(self, organization: Organization) -> None:
-        """Unvalidated VAT ID should fail."""
-        org = _make_org_invoicing_ready(organization)
-        org.vat_id_validated = False
-        org.save()
-        with pytest.raises(HttpError) as exc_info:
-            validate_invoicing_prerequisites(org)
-        assert "VIES" in str(exc_info.value)
-
-    def test_fails_when_billing_name_missing(self, organization: Organization) -> None:
-        """Missing billing name should fail."""
-        org = _make_org_invoicing_ready(organization)
-        org.billing_name = ""
-        org.save()
-        with pytest.raises(HttpError) as exc_info:
-            validate_invoicing_prerequisites(org)
-        assert "Billing name" in str(exc_info.value)
-
-    def test_fails_when_billing_address_missing(self, organization: Organization) -> None:
-        """Missing billing address should fail."""
-        org = _make_org_invoicing_ready(organization)
-        org.billing_address = ""
-        org.save()
-        with pytest.raises(HttpError) as exc_info:
-            validate_invoicing_prerequisites(org)
-        assert "Billing address" in str(exc_info.value)
-
-    def test_reports_all_missing_fields_at_once(self, organization: Organization) -> None:
-        """All missing prerequisites should be reported in a single error."""
-        with pytest.raises(HttpError) as exc_info:
-            validate_invoicing_prerequisites(organization)
-        msg = str(exc_info.value)
-        for expected in ("EU-based", "VIES", "Billing name", "Billing address"):
-            assert expected in msg
-
-
-# ---------------------------------------------------------------------------
-# set_invoicing_mode
-# ---------------------------------------------------------------------------
-
-
-class TestSetInvoicingMode:
-    """Test invoicing mode transitions."""
-
-    def test_set_none_always_allowed(self, organization: Organization) -> None:
-        """Setting NONE requires no prerequisites."""
-        result = set_invoicing_mode(organization, Organization.InvoicingMode.NONE)
-        assert result.invoicing_mode == Organization.InvoicingMode.NONE
-
-    def test_set_hybrid_requires_prerequisites(self, organization: Organization) -> None:
-        """Setting HYBRID without prerequisites should raise 422."""
-        with pytest.raises(HttpError) as exc_info:
-            set_invoicing_mode(organization, Organization.InvoicingMode.HYBRID)
-        assert exc_info.value.status_code == 422
-
-    def test_set_auto_requires_prerequisites(self, organization: Organization) -> None:
-        """Setting AUTO without prerequisites should raise 422."""
-        with pytest.raises(HttpError):
-            set_invoicing_mode(organization, Organization.InvoicingMode.AUTO)
-
-    def test_set_hybrid_with_prerequisites(self, organization: Organization) -> None:
-        """Setting HYBRID succeeds when org meets all prerequisites."""
-        org = _make_org_invoicing_ready(organization)
-        result = set_invoicing_mode(org, Organization.InvoicingMode.HYBRID)
-        assert result.invoicing_mode == Organization.InvoicingMode.HYBRID
-        org.refresh_from_db()
-        assert org.invoicing_mode == Organization.InvoicingMode.HYBRID
-
-    def test_set_auto_with_prerequisites(self, organization: Organization) -> None:
-        """Setting AUTO succeeds when org meets all prerequisites."""
-        org = _make_org_invoicing_ready(organization)
-        result = set_invoicing_mode(org, Organization.InvoicingMode.AUTO)
-        assert result.invoicing_mode == Organization.InvoicingMode.AUTO
-
-    def test_downgrade_from_auto_to_none(self, organization: Organization) -> None:
-        """Downgrading from AUTO to NONE should always work."""
-        org = _make_org_invoicing_ready(organization)
-        set_invoicing_mode(org, Organization.InvoicingMode.AUTO)
-        result = set_invoicing_mode(org, Organization.InvoicingMode.NONE)
-        assert result.invoicing_mode == Organization.InvoicingMode.NONE
-
-
-# ---------------------------------------------------------------------------
-# generate_attendee_invoice
-# ---------------------------------------------------------------------------
-
-
-class TestGenerateAttendeeInvoice:
-    """Test invoice generation for completed checkout sessions."""
-
-    @patch(MOCK_RENDER_PDF, return_value=b"fake-pdf")
-    def test_hybrid_mode_creates_draft(
-        self,
-        mock_pdf: t.Any,
-        organization: Organization,
-        event: Event,
-        event_ticket_tier: TicketTier,
-        member_user: RevelUser,
-    ) -> None:
-        """HYBRID mode should create an invoice with DRAFT status."""
-        org = _make_org_invoicing_ready(organization)
-        org.invoicing_mode = Organization.InvoicingMode.HYBRID
-        org.save()
-        _create_payment(
-            user=member_user,
-            event=event,
-            tier=event_ticket_tier,
-            session_id="cs_hybrid_1",
-            buyer_billing_snapshot=_default_billing_snapshot(),
-        )
-        invoice = generate_attendee_invoice("cs_hybrid_1")
-        assert invoice is not None
-        assert invoice.status == AttendeeInvoice.InvoiceStatus.DRAFT
-        assert invoice.issued_at is None
-        assert invoice.organization == org
-        assert invoice.user == member_user
-
-    @patch(MOCK_RENDER_PDF, return_value=b"fake-pdf")
-    def test_auto_mode_creates_issued_invoice(
-        self,
-        mock_pdf: t.Any,
-        organization: Organization,
-        event: Event,
-        event_ticket_tier: TicketTier,
-        member_user: RevelUser,
-    ) -> None:
-        """AUTO mode should create an invoice with ISSUED status and issued_at set."""
-        org = _make_org_invoicing_ready(organization)
-        org.invoicing_mode = Organization.InvoicingMode.AUTO
-        org.save()
-        _create_payment(
-            user=member_user,
-            event=event,
-            tier=event_ticket_tier,
-            session_id="cs_auto_1",
-            buyer_billing_snapshot=_default_billing_snapshot(),
-        )
-        invoice = generate_attendee_invoice("cs_auto_1")
-        assert invoice is not None
-        assert invoice.status == AttendeeInvoice.InvoiceStatus.ISSUED
-        assert invoice.issued_at is not None
-
-    @patch(MOCK_RENDER_PDF, return_value=b"fake-pdf")
-    def test_no_billing_snapshot_returns_none(
-        self,
-        mock_pdf: t.Any,
-        organization: Organization,
-        event: Event,
-        event_ticket_tier: TicketTier,
-        member_user: RevelUser,
-    ) -> None:
-        """No buyer billing snapshot should return None."""
-        org = _make_org_invoicing_ready(organization)
-        org.invoicing_mode = Organization.InvoicingMode.AUTO
-        org.save()
-        _create_payment(
-            user=member_user,
-            event=event,
-            tier=event_ticket_tier,
-            session_id="cs_no_billing",
-            buyer_billing_snapshot=None,
-        )
-        assert generate_attendee_invoice("cs_no_billing") is None
-
-    def test_org_invoicing_mode_none_returns_none(
-        self,
-        organization: Organization,
-        event: Event,
-        event_ticket_tier: TicketTier,
-        member_user: RevelUser,
-    ) -> None:
-        """Org with invoicing_mode=NONE should not generate an invoice."""
-        _create_payment(
-            user=member_user,
-            event=event,
-            tier=event_ticket_tier,
-            session_id="cs_no_inv",
-            buyer_billing_snapshot=_default_billing_snapshot(),
-        )
-        assert generate_attendee_invoice("cs_no_inv") is None
-
-    def test_no_payments_returns_none(self) -> None:
-        """No payments for the session should return None."""
-        assert generate_attendee_invoice("cs_nonexistent") is None
-
-    @patch(MOCK_RENDER_PDF, return_value=b"fake-pdf")
-    def test_idempotency_returns_same_invoice(
-        self,
-        mock_pdf: t.Any,
-        organization: Organization,
-        event: Event,
-        event_ticket_tier: TicketTier,
-        member_user: RevelUser,
-    ) -> None:
-        """Calling generate twice for the same session returns the same invoice."""
-        org = _make_org_invoicing_ready(organization)
-        org.invoicing_mode = Organization.InvoicingMode.HYBRID
-        org.save()
-        _create_payment(
-            user=member_user,
-            event=event,
-            tier=event_ticket_tier,
-            session_id="cs_idemp",
-            buyer_billing_snapshot=_default_billing_snapshot(),
-        )
-        inv1 = generate_attendee_invoice("cs_idemp")
-        inv2 = generate_attendee_invoice("cs_idemp")
-        assert inv1 is not None and inv2 is not None
-        assert inv1.id == inv2.id
-
-    @patch(MOCK_RENDER_PDF, return_value=b"fake-pdf")
-    def test_line_items_built_correctly(
-        self,
-        mock_pdf: t.Any,
-        organization: Organization,
-        event: Event,
-        event_ticket_tier: TicketTier,
-        member_user: RevelUser,
-    ) -> None:
-        """Line items should contain event name, tier name, and guest name."""
-        org = _make_org_invoicing_ready(organization)
-        org.invoicing_mode = Organization.InvoicingMode.HYBRID
-        org.save()
-        _create_payment(
-            user=member_user,
-            event=event,
-            tier=event_ticket_tier,
-            session_id="cs_li",
-            buyer_billing_snapshot=_default_billing_snapshot(),
-            guest_name="John Doe",
-        )
-        invoice = generate_attendee_invoice("cs_li")
-        assert invoice is not None
-        assert len(invoice.line_items) == 1
-        item = invoice.line_items[0]
-        assert event.name in item["description"]
-        assert event_ticket_tier.name in item["description"]
-        assert "John Doe" in item["description"]
-
-    @patch(MOCK_RENDER_PDF, return_value=b"fake-pdf")
-    def test_line_item_description_omits_blank_holder_name(
-        self,
-        mock_pdf: t.Any,
-        organization: Organization,
-        event: Event,
-        event_ticket_tier: TicketTier,
-        member_user: RevelUser,
-    ) -> None:
-        """A blank holder name (#845) must not leave a trailing separator on the invoice line."""
-        org = _make_org_invoicing_ready(organization)
-        org.invoicing_mode = Organization.InvoicingMode.HYBRID
-        org.save()
-        _create_payment(
-            user=member_user,
-            event=event,
-            tier=event_ticket_tier,
-            session_id="cs_li_blank",
-            buyer_billing_snapshot=_default_billing_snapshot(),
-            guest_name="",
-        )
-        invoice = generate_attendee_invoice("cs_li_blank")
-        assert invoice is not None
-        assert invoice.line_items[0]["description"] == f"{event.name} — {event_ticket_tier.name}"
-
-    @patch(MOCK_RENDER_PDF, return_value=b"fake-pdf")
-    def test_seller_and_buyer_snapshots(
-        self,
-        mock_pdf: t.Any,
-        organization: Organization,
-        event: Event,
-        event_ticket_tier: TicketTier,
-        member_user: RevelUser,
-    ) -> None:
-        """Invoice should snapshot seller (org) and buyer info correctly."""
-        org = _make_org_invoicing_ready(organization)
-        org.invoicing_mode = Organization.InvoicingMode.HYBRID
-        org.save()
-        snapshot = _default_billing_snapshot()
-        _create_payment(
-            user=member_user, event=event, tier=event_ticket_tier, session_id="cs_snap", buyer_billing_snapshot=snapshot
-        )
-        invoice = generate_attendee_invoice("cs_snap")
-        assert invoice is not None
-        assert invoice.seller_name == org.billing_name
-        assert invoice.seller_vat_id == org.vat_id
-        assert invoice.buyer_name == snapshot["billing_name"]
-        assert invoice.buyer_email == snapshot["billing_email"]
-
-    @patch(MOCK_RENDER_PDF, return_value=b"fake-pdf")
-    def test_reverse_charge_detected(
-        self,
-        mock_pdf: t.Any,
-        organization: Organization,
-        event: Event,
-        event_ticket_tier: TicketTier,
-        member_user: RevelUser,
-    ) -> None:
-        """Reverse charge flag set when billing snapshot records reverse_charge=True.
-
-        Snapshot passthrough only: checkout no longer produces reverse_charge=True
-        snapshots (#868 — admission is taxed where the event takes place), but the
-        invoice service still honors historical snapshots on webhook replays.
-        """
-        org = _make_org_invoicing_ready(organization)
-        org.invoicing_mode = Organization.InvoicingMode.HYBRID
-        org.save()
-        _create_payment(
-            user=member_user,
-            event=event,
-            tier=event_ticket_tier,
-            session_id="cs_rc",
-            amount=Decimal("81.97"),
-            net_amount=Decimal("81.97"),
-            vat_amount=Decimal("0.00"),
-            vat_rate=Decimal("0.00"),
-            buyer_billing_snapshot=_default_billing_snapshot(reverse_charge=True),
-        )
-        invoice = generate_attendee_invoice("cs_rc")
-        assert invoice is not None
-        assert invoice.reverse_charge is True
-        assert invoice.total_vat == Decimal("0.00")
-
-    @patch(MOCK_RENDER_PDF, return_value=b"fake-pdf")
-    def test_totals_aggregated_from_multiple_payments(
-        self,
-        mock_pdf: t.Any,
-        organization: Organization,
-        event: Event,
-        event_ticket_tier: TicketTier,
-        member_user: RevelUser,
-    ) -> None:
-        """Invoice totals should be aggregated from all payments in the session."""
-        org = _make_org_invoicing_ready(organization)
-        org.invoicing_mode = Organization.InvoicingMode.HYBRID
-        org.save()
-        for i in range(2):
-            _create_payment(
-                user=member_user,
-                event=event,
-                tier=event_ticket_tier,
-                session_id="cs_multi",
-                buyer_billing_snapshot=_default_billing_snapshot(),
-                guest_name=f"Guest {i}",
-            )
-        invoice = generate_attendee_invoice("cs_multi")
-        assert invoice is not None
-        assert invoice.total_gross == Decimal("200.00")
-        assert invoice.total_net == Decimal("163.94")
-        assert len(invoice.line_items) == 2
-
-
-# ---------------------------------------------------------------------------
-# _is_export — PDF export wording gate
-# ---------------------------------------------------------------------------
-
-
-class TestIsExport:
-    """Export wording is gated on total_vat == 0 (#868).
-
-    Admission is never zero-rated as an export anymore, so only historical
-    zero-rated documents (which must re-render as issued) get the export
-    wording; new non-EU invoices carry full VAT and render normally.
-    """
-
-    @staticmethod
-    def _invoice(**overrides: t.Any) -> AttendeeInvoice:
-        """An unsaved invoice shaped like a historical non-EU zero-rated document."""
-        fields: dict[str, t.Any] = {
-            "buyer_vat_country": "US",
-            "reverse_charge": False,
-            "total_vat": Decimal("0.00"),
-        }
-        fields.update(overrides)
-        return AttendeeInvoice(**fields)
-
-    def test_non_eu_invoice_with_vat_is_not_export(self) -> None:
-        """A new non-EU invoice carrying full VAT (#868) must not render export wording."""
-        assert _is_export(self._invoice(total_vat=Decimal("18.03"))) is False
-
-    def test_historical_zero_rated_non_eu_invoice_is_export(self) -> None:
-        """A historical document (total_vat=0, non-EU buyer, not RC) still renders as export."""
-        assert _is_export(self._invoice()) is True
-
-    def test_reverse_charge_invoice_is_not_export(self) -> None:
-        """Reverse-charge documents get RC wording, never export wording."""
-        assert _is_export(self._invoice(reverse_charge=True)) is False
-
-    def test_eu_buyer_zero_vat_invoice_is_not_export(self) -> None:
-        """A zero-VAT invoice for an EU buyer (e.g. 0% org rate) is not an export."""
-        assert _is_export(self._invoice(buyer_vat_country="DE")) is False
-
-
-# ---------------------------------------------------------------------------
-# update_draft_invoice
-# ---------------------------------------------------------------------------
-
-
-class TestUpdateDraftInvoice:
-    """Test editing draft invoices."""
-
-    def test_can_edit_buyer_name(self, organization: Organization, event: Event, member_user: RevelUser) -> None:
-        """Editing buyer_name on a draft should succeed."""
-        inv = _create_draft(organization, event, member_user)
-        assert update_draft_invoice(inv, {"buyer_name": "New Name"}).buyer_name == "New Name"
-
-    @pytest.mark.parametrize("field", DERIVED_TOTAL_FIELDS)
-    def test_cannot_edit_derived_total(
-        self, field: str, organization: Organization, event: Event, member_user: RevelUser
-    ) -> None:
-        """Regression (#911): a header total cannot be stated independently of the lines.
-
-        Setting ``total_gross`` alone used to succeed, leaving a header claiming
-        one amount above line items summing to another -- a contradiction the
-        API returns in a single response now that ``vat_breakdown`` is exposed
-        (#910), and that issuance then turned into a delivered PDF.
-        """
-        inv = _create_draft(organization, event, member_user)
-        with pytest.raises(HttpError) as exc_info:
-            update_draft_invoice(inv, {field: Decimal("200.00")})
-        assert exc_info.value.status_code == 422
-        assert field in str(exc_info.value)
-
-    def test_editing_line_items_recomputes_totals(
-        self, organization: Organization, event: Event, member_user: RevelUser
-    ) -> None:
-        """Rewriting the lines rewrites the header they summarize (#911)."""
-        inv = _create_draft(organization, event, member_user)
-        updated = update_draft_invoice(
-            inv,
-            {
-                "line_items": [
-                    {
-                        "description": "Ticket A",
-                        "unit_price_gross": Decimal("50.00"),
-                        "discount_amount": Decimal("5.00"),
-                        "net_amount": Decimal("40.98"),
-                        "vat_amount": Decimal("9.02"),
-                        "vat_rate": Decimal("22.00"),
-                    },
-                    {
-                        "description": "Ticket B",
-                        "unit_price_gross": Decimal("30.00"),
-                        "discount_amount": Decimal("0.00"),
-                        "net_amount": Decimal("27.27"),
-                        "vat_amount": Decimal("2.73"),
-                        "vat_rate": Decimal("10.00"),
-                    },
-                ]
-            },
-        )
-        updated.refresh_from_db()
-        assert updated.total_gross == Decimal("80.00")
-        assert updated.total_net == Decimal("68.25")
-        assert updated.total_vat == Decimal("11.75")
-        assert updated.discount_amount_total == Decimal("5.00")
-        # The scalar header rate follows generation's convention: the first line's.
-        assert updated.vat_rate == Decimal("22.00")
-
-    def test_recomputed_totals_reconcile_with_vat_breakdown(
-        self, organization: Organization, event: Event, member_user: RevelUser
-    ) -> None:
-        """The invariant ``InvoiceVatBucketDict`` promises holds after an edit (#911)."""
-        inv = _create_draft(organization, event, member_user)
-        updated = update_draft_invoice(
-            inv,
-            {
-                "line_items": [
-                    {
-                        "description": "Standard",
-                        "unit_price_gross": Decimal("111.00"),
-                        "discount_amount": Decimal("0.00"),
-                        "net_amount": Decimal("91.74"),
-                        "vat_amount": Decimal("19.26"),
-                        "vat_rate": Decimal("21.00"),
-                    },
-                    {
-                        "description": "Reduced",
-                        "unit_price_gross": Decimal("22.00"),
-                        "discount_amount": Decimal("0.00"),
-                        "net_amount": Decimal("20.00"),
-                        "vat_amount": Decimal("2.00"),
-                        "vat_rate": Decimal("10.00"),
-                    },
-                ]
-            },
-        )
-        buckets = updated.vat_breakdown
-        assert updated.has_mixed_vat_rates
-        assert sum(b["gross_amount"] for b in buckets) == updated.total_gross
-        assert sum(b["net_amount"] for b in buckets) == updated.total_net
-        assert sum(b["vat_amount"] for b in buckets) == updated.total_vat
-
-    def test_empty_update_returns_unchanged(
-        self, organization: Organization, event: Event, member_user: RevelUser
-    ) -> None:
-        """Empty update data should return the invoice unchanged."""
-        inv = _create_draft(organization, event, member_user)
-        assert update_draft_invoice(inv, {}).buyer_name == "Original Buyer"
-
-    def test_cannot_edit_issued_invoice(self, organization: Organization, event: Event, member_user: RevelUser) -> None:
-        """Attempting to edit an ISSUED invoice should raise 409."""
-        inv = _create_issued(organization, event, member_user)
-        with pytest.raises(HttpError) as exc_info:
-            update_draft_invoice(inv, {"buyer_name": "Hacked"})
-        assert exc_info.value.status_code == 409
-
-    def test_cannot_edit_disallowed_fields(
-        self, organization: Organization, event: Event, member_user: RevelUser
-    ) -> None:
-        """Attempting to edit seller_name should raise 422."""
-        inv = _create_draft(organization, event, member_user)
-        with pytest.raises(HttpError) as exc_info:
-            update_draft_invoice(inv, {"seller_name": "Hacked"})
-        assert exc_info.value.status_code == 422
-        assert "seller_name" in str(exc_info.value)
-
-    @patch(MOCK_RENDER_PDF, return_value=b"fake-pdf")
-    def test_edit_invalidates_stale_pdf(
-        self,
-        mock_pdf: t.Any,
-        organization: Organization,
-        event: Event,
-        event_ticket_tier: TicketTier,
-        member_user: RevelUser,
-    ) -> None:
-        """Editing a draft with an existing PDF should delete the stale PDF."""
-        org = _make_org_invoicing_ready(organization)
-        org.invoicing_mode = Organization.InvoicingMode.HYBRID
-        org.save()
-        _create_payment(
-            user=member_user,
-            event=event,
-            tier=event_ticket_tier,
-            session_id="cs_stale",
-            buyer_billing_snapshot=_default_billing_snapshot(),
-        )
-        invoice = generate_attendee_invoice("cs_stale")
-        assert invoice is not None and invoice.pdf_file
-        update_draft_invoice(invoice, {"buyer_name": "Updated"})
-        invoice.refresh_from_db()
-        assert not invoice.pdf_file
-
-    def test_all_editable_fields_accepted(
-        self, organization: Organization, event: Event, member_user: RevelUser
-    ) -> None:
-        """All fields in EDITABLE_DRAFT_FIELDS should be accepted."""
-        inv = _create_draft(organization, event, member_user)
-        data: dict[str, t.Any] = {
-            "buyer_name": "Updated Buyer",
-            "buyer_vat_id": "DE123456789",
-            "buyer_vat_country": "DE",
-            "buyer_address": "Berlin",
-            "buyer_email": "test@example.de",
-            "currency": "USD",
-            "reverse_charge": True,
-            "discount_code_text": "SALE10",
-            "line_items": [{"description": "Test"}],
-        }
-        update_draft_invoice(inv, data)  # Should not raise
-
-    def test_edit_rejects_an_invoice_issued_since_it_was_fetched(
-        self, organization: Organization, event: Event, member_user: RevelUser
-    ) -> None:
-        """The DRAFT check must judge the locked row, not the caller's snapshot (#911 review).
-
-        The controller fetches the invoice, then hands it here. Unlocked, a
-        concurrent issue landing in that window left this function checking a
-        stale DRAFT and editing an already-issued legal document.
-        """
-        inv = _create_draft(organization, event, member_user)
-        stale = AttendeeInvoice.objects.get(pk=inv.pk)  # what the controller handed us
-        AttendeeInvoice.objects.filter(pk=inv.pk).update(status=AttendeeInvoice.InvoiceStatus.ISSUED)
-
-        with pytest.raises(HttpError) as exc_info:
-            update_draft_invoice(stale, {"buyer_name": "Too late"})
-
-        assert exc_info.value.status_code == 409
-        inv.refresh_from_db()
-        assert inv.buyer_name == "Original Buyer"
-
-    @patch(MOCK_RENDER_PDF, return_value=b"fake-pdf")
-    def test_issue_reconciles_the_locked_row_not_the_callers_snapshot(
-        self,
-        mock_pdf: t.Any,
-        organization: Organization,
-        event: Event,
-        member_user: RevelUser,
-    ) -> None:
-        """Reconciliation must judge the committed row (#911 review).
-
-        Unlocked, a PATCH landing after the controller's fetch was validated in
-        absentia: the guard passed on the stale copy and the PDF rendered from it,
-        so the buyer received a document contradicting the stored invoice.
-        """
-        inv = _create_draft(organization, event, member_user)
-        stale = AttendeeInvoice.objects.get(pk=inv.pk)
-        AttendeeInvoice.objects.filter(pk=inv.pk).update(total_gross=Decimal("999.00"))
-
-        with pytest.raises(HttpError) as exc_info:
-            issue_draft_invoice(stale)
-
-        assert exc_info.value.status_code == 422
-        inv.refresh_from_db()
-        assert inv.status == AttendeeInvoice.InvoiceStatus.DRAFT
-        mock_pdf.assert_not_called()
-
-    @pytest.mark.parametrize(
-        "field",
-        ["buyer_vat_id", "buyer_vat_country", "buyer_address", "buyer_email", "discount_code_text"],
-    )
-    def test_null_blankable_field_coerced_to_empty_string(
-        self, field: str, organization: Organization, event: Event, member_user: RevelUser
-    ) -> None:
-        """Regression (#700): null on a blankable (blank=True) column is coerced to "" (NOT NULL)."""
-        inv = _create_draft(organization, event, member_user)
-        updated = update_draft_invoice(inv, {field: None})
-        assert getattr(updated, field) == ""
-        updated.refresh_from_db()
-        assert getattr(updated, field) == ""
-
-    @pytest.mark.parametrize("field", ["buyer_name", "currency"])
-    def test_null_required_field_rejected(
-        self, field: str, organization: Organization, event: Event, member_user: RevelUser
-    ) -> None:
-        """Required columns lack blank=True, so an explicit null is rejected up front (#908).
-
-        Previously this reached ``full_clean()`` and surfaced as a Django
-        ``ValidationError`` (a 400); the null guard now rejects it as a 422
-        naming the offending field, before anything is written.
-        """
-        inv = _create_draft(organization, event, member_user)
-        with pytest.raises(HttpError) as exc_info:
-            update_draft_invoice(inv, {field: None})
-        assert exc_info.value.status_code == 422
-        assert field in str(exc_info.value)
-
-
-# ---------------------------------------------------------------------------
-# issue_draft_invoice
-# ---------------------------------------------------------------------------
-
-
-class TestIssueDraftInvoice:
-    """Test issuing a draft invoice."""
-
-    @patch(MOCK_RENDER_PDF, return_value=b"fake-pdf")
-    def test_transitions_to_issued(
-        self,
-        mock_pdf: t.Any,
-        organization: Organization,
-        event: Event,
-        member_user: RevelUser,
-    ) -> None:
-        """Issuing a draft should set status=ISSUED and issued_at."""
-        inv = _create_draft(organization, event, member_user)
-        result = issue_draft_invoice(inv)
-        assert result.status == AttendeeInvoice.InvoiceStatus.ISSUED
-        assert result.issued_at is not None
-        mock_pdf.assert_called_once()
-
-    @patch(MOCK_RENDER_PDF, return_value=b"fake-pdf")
-    def test_reissue_already_issued_is_idempotent(
-        self,
-        mock_pdf: t.Any,
-        organization: Organization,
-        event: Event,
-        member_user: RevelUser,
-    ) -> None:
-        """Re-issuing an already-issued invoice should succeed (regenerate PDF)."""
-        inv = _create_issued(organization, event, member_user)
-        original_issued_at = inv.issued_at
-        result = issue_draft_invoice(inv)
-        assert result.status == AttendeeInvoice.InvoiceStatus.ISSUED
-        assert result.issued_at == original_issued_at  # not re-set
-        mock_pdf.assert_called_once()
-
-    def test_cannot_issue_cancelled(self, organization: Organization, event: Event, member_user: RevelUser) -> None:
-        """Issuing a cancelled invoice should raise 409."""
-        inv = _create_draft(organization, event, member_user)
-        inv.status = AttendeeInvoice.InvoiceStatus.CANCELLED
-        inv.save()
-        with pytest.raises(HttpError) as exc_info:
-            issue_draft_invoice(inv)
-        assert exc_info.value.status_code == 409
-
-    @patch(MOCK_RENDER_PDF, return_value=b"fake-pdf")
-    def test_cannot_issue_draft_whose_totals_drifted(
-        self,
-        mock_pdf: t.Any,
-        organization: Organization,
-        event: Event,
-        member_user: RevelUser,
-    ) -> None:
-        """Regression (#911): a drifted DRAFT must not become a legal document.
-
-        The API can no longer produce one, but a row written straight through
-        the Django admin -- or one that drifted before #911 -- still can, and
-        issuance is its own gate: it used to return 200 and deliver a PDF whose
-        header contradicted its own line items.
-        """
-        inv = _create_draft(organization, event, member_user)
-        AttendeeInvoice.objects.filter(pk=inv.pk).update(total_gross=Decimal("999.00"))
-        inv.refresh_from_db()
-
-        with pytest.raises(HttpError) as exc_info:
-            issue_draft_invoice(inv)
-
-        assert exc_info.value.status_code == 422
-        assert "total_gross" in str(exc_info.value)
-        assert "999.00" in str(exc_info.value)
-        inv.refresh_from_db()
-        assert inv.status == AttendeeInvoice.InvoiceStatus.DRAFT
-        mock_pdf.assert_not_called()
-
-    @patch(MOCK_RENDER_PDF, return_value=b"fake-pdf")
-    def test_issue_guard_tolerates_a_line_item_without_discount_amount(
-        self,
-        mock_pdf: t.Any,
-        organization: Organization,
-        event: Event,
-        member_user: RevelUser,
-    ) -> None:
-        """The guard must judge a row, not crash on it (#911 review).
-
-        ``discount_amount`` is the one key ``vat_breakdown`` never reads, so a
-        row lacking it renders fine everywhere else. The guard runs on rows
-        nobody normalized, where a bare ``item["discount_amount"]`` would have
-        been a 500 on exactly the malformed document it exists to catch.
-        """
-        inv = _create_draft(organization, event, member_user)
-        AttendeeInvoice.objects.filter(pk=inv.pk).update(
-            line_items=[
-                {
-                    "description": "legacy row",
-                    "unit_price_gross": "100.00",
-                    "net_amount": "81.97",
-                    "vat_amount": "18.03",
-                    "vat_rate": "22.00",
-                }
-            ]
-        )
-        inv.refresh_from_db()
-
-        assert inv.vat_breakdown  # readable, as it always was
-        assert issue_draft_invoice(inv).status == AttendeeInvoice.InvoiceStatus.ISSUED
-
-    @patch(MOCK_RENDER_PDF, return_value=b"fake-pdf")
-    def test_patching_line_items_repairs_a_drifted_draft(
-        self,
-        mock_pdf: t.Any,
-        organization: Organization,
-        event: Event,
-        member_user: RevelUser,
-    ) -> None:
-        """The documented recovery path: restate the lines, the header follows (#911)."""
-        inv = _create_draft(organization, event, member_user)
-        AttendeeInvoice.objects.filter(pk=inv.pk).update(total_gross=Decimal("999.00"))
-        inv.refresh_from_db()
-
-        repaired = update_draft_invoice(inv, {"line_items": list(inv.line_items)})
-
-        assert repaired.total_gross == Decimal("100.00")
-        assert issue_draft_invoice(repaired).status == AttendeeInvoice.InvoiceStatus.ISSUED
-
-    @patch(MOCK_RENDER_PDF, return_value=b"fake-pdf")
-    def test_reissue_of_drifted_issued_invoice_still_regenerates_pdf(
-        self,
-        mock_pdf: t.Any,
-        organization: Organization,
-        event: Event,
-        member_user: RevelUser,
-    ) -> None:
-        """The guard covers only DRAFT -> ISSUED (#911).
-
-        Re-issuing exists to recover a failed PDF generation on an already-issued
-        document; refusing it would strand a row that drifted before the guard
-        landed with no PDF at all.
-        """
-        inv = _create_issued(organization, event, member_user)
-        AttendeeInvoice.objects.filter(pk=inv.pk).update(total_gross=Decimal("999.00"))
-        inv.refresh_from_db()
-
-        assert issue_draft_invoice(inv).status == AttendeeInvoice.InvoiceStatus.ISSUED
-        mock_pdf.assert_called_once()
-
-
-# ---------------------------------------------------------------------------
-# delete_draft_invoice
-# ---------------------------------------------------------------------------
-
-
-class TestDeleteDraftInvoice:
-    """Test draft invoice deletion."""
-
-    def test_can_delete_draft(self, organization: Organization, event: Event, member_user: RevelUser) -> None:
-        """Deleting a draft invoice should remove it from the database."""
-        inv = _create_draft(organization, event, member_user)
-        inv_id = inv.id
-        delete_draft_invoice(inv)
-        assert not AttendeeInvoice.objects.filter(id=inv_id).exists()
-
-    def test_cannot_delete_issued(self, organization: Organization, event: Event, member_user: RevelUser) -> None:
-        """Deleting an issued invoice should raise 409."""
-        inv = _create_issued(organization, event, member_user)
-        with pytest.raises(HttpError) as exc_info:
-            delete_draft_invoice(inv)
-        assert exc_info.value.status_code == 409
