--- src/events/tests/test_service/test_organization_service.py
@@ -1,985 +1,0 @@
-"""Tests for the organization service."""
-
-import typing as t
-from contextlib import contextmanager
-from unittest.mock import MagicMock, patch
-
-import pytest
-from django.conf import settings
-from django.core.exceptions import ValidationError
-from django.utils import timezone
-from ninja.errors import HttpError
-
-from accounts.jwt import blacklist as blacklist_token
-from accounts.jwt import create_token
-from accounts.models import RevelUser
-from events import schema
-from events.exceptions import (
-    AlreadyMemberError,
-    OrganizationTokenGrantInvariantError,
-    OrganizationTokenMembershipTierRequiredError,
-    OrganizationTokenStaffGrantForbidden,
-    PendingMembershipRequestExistsError,
-)
-from events.models import (
-    MembershipTier,
-    Organization,
-    OrganizationMember,
-    OrganizationMembershipRequest,
-    OrganizationStaff,
-    OrganizationToken,
-    PermissionMap,
-    PermissionsSchema,
-)
-from events.service import organization_service
-
-
-@contextmanager
-def _force_first_lookup_miss(manager: t.Any) -> t.Iterator[None]:
-    """Make ``manager.filter(...)`` miss on its first call, then behave normally.
-
-    Simulates the concurrent double-submit the partial unique constraint now
-    catches: the racing row is already committed by the time our INSERT lands,
-    but our own lookup ran too early to see it.
-    """
-    call_count = 0
-    real_filter = manager.filter
-
-    def fake_filter(*args: t.Any, **kwargs: t.Any) -> t.Any:
-        nonlocal call_count
-        call_count += 1
-        if call_count == 1:
-            return manager.none()
-        return real_filter(*args, **kwargs)
-
-    with patch.object(manager, "filter", side_effect=fake_filter):
-        yield
-
-
-@pytest.fixture
-def organization_token(organization: Organization, organization_owner_user: RevelUser) -> OrganizationToken:
-    """An organization token that grants membership."""
-    from events.models import MembershipTier
-
-    default_tier = MembershipTier.objects.get(organization=organization, name="General membership")
-    return OrganizationToken.objects.create(
-        organization=organization, issuer=organization_owner_user, membership_tier=default_tier
-    )
-
-
-@pytest.fixture
-def staff_organization_token(organization: Organization, organization_owner_user: RevelUser) -> OrganizationToken:
-    """An organization token that grants staff permissions."""
-    return OrganizationToken.objects.create(
-        organization=organization, issuer=organization_owner_user, grants_staff_status=True, grants_membership=False
-    )
-
-
-@pytest.mark.django_db
-class TestCreateMembershipRequest:
-    """Tests for the create_membership_request function."""
-
-    def test_create_membership_request_success(self, organization: Organization, nonmember_user: RevelUser) -> None:
-        """Test that a membership request is created successfully."""
-        # Act
-        request = organization_service.create_membership_request(organization, nonmember_user)
-
-        # Assert
-        assert OrganizationMembershipRequest.objects.filter(organization=organization, user=nonmember_user).exists()
-        assert request.status == OrganizationMembershipRequest.Status.PENDING
-
-    def test_create_membership_request_already_member_fails(self, organization_membership: OrganizationMember) -> None:
-        """Test that a membership request is not created if the user is already a member."""
-        # Act & Assert
-        with pytest.raises(AlreadyMemberError):
-            organization_service.create_membership_request(
-                organization_membership.organization, organization_membership.user
-            )
-
-    def test_create_membership_request_pending_request_exists_fails(
-        self, organization: Organization, nonmember_user: RevelUser
-    ) -> None:
-        """Test that a membership request is not created if a pending request already exists."""
-        # Arrange
-        OrganizationMembershipRequest.objects.create(organization=organization, user=nonmember_user)
-
-        # Act & Assert
-        with pytest.raises(PendingMembershipRequestExistsError):
-            organization_service.create_membership_request(organization, nonmember_user)
-
-    def test_create_membership_request_blacklisted_user_fails(
-        self, organization: Organization, nonmember_user: RevelUser, organization_owner_user: RevelUser
-    ) -> None:
-        """Test that a blacklisted user cannot create a membership request."""
-        from events.models import Blacklist
-
-        # Arrange - blacklist the user by direct FK match
-        Blacklist.objects.create(
-            organization=organization,
-            user=nonmember_user,
-            email=nonmember_user.email,
-            created_by=organization_owner_user,
-            reason="Test blacklist",
-        )
-
-        # Act & Assert
-        with pytest.raises(HttpError) as exc_info:
-            organization_service.create_membership_request(organization, nonmember_user)
-        assert exc_info.value.status_code == 403
-
-    def test_create_membership_request_blacklisted_by_email_fails(
-        self, organization: Organization, nonmember_user: RevelUser, organization_owner_user: RevelUser
-    ) -> None:
-        """Test that a user blacklisted by email (without FK) cannot create a membership request."""
-        from events.models import Blacklist
-
-        # Arrange - blacklist by email only (no user FK)
-        Blacklist.objects.create(
-            organization=organization,
-            email=nonmember_user.email,
-            created_by=organization_owner_user,
-            reason="Test blacklist by email",
-        )
-
-        # Act & Assert
-        with pytest.raises(HttpError) as exc_info:
-            organization_service.create_membership_request(organization, nonmember_user)
-        assert exc_info.value.status_code == 403
-
-    def test_create_membership_request_lost_race_raises_domain_error(
-        self, organization: Organization, nonmember_user: RevelUser
-    ) -> None:
-        """A concurrent double-submit must still yield the 409 domain error, not a 500.
-
-        ``unique_pending_application_per_user_org_tier`` rejects the second insert.
-        This is the ValidationError arm: ``TimeStampedModel.save`` runs ``full_clean``,
-        so ``validate_constraints`` sees the committed racing row before the INSERT.
-        """
-        existing = OrganizationMembershipRequest.objects.create(organization=organization, user=nonmember_user)
-
-        with _force_first_lookup_miss(OrganizationMembershipRequest.objects):
-            with pytest.raises(PendingMembershipRequestExistsError):
-                organization_service.create_membership_request(organization, nonmember_user)
-
-        surviving = OrganizationMembershipRequest.objects.filter(organization=organization, user=nonmember_user)
-        assert [row.pk for row in surviving] == [existing.pk]
-
-    def test_create_membership_request_lost_db_level_race_raises_domain_error(
-        self, organization: Organization, nonmember_user: RevelUser
-    ) -> None:
-        """The IntegrityError arm: ``full_clean`` disabled, so the unique index rejects the INSERT.
-
-        Also pins that the ambient transaction survives — the helper's savepoint is
-        what keeps the recovery re-fetch from raising ``TransactionManagementError``.
-        """
-        existing = OrganizationMembershipRequest.objects.create(organization=organization, user=nonmember_user)
-
-        with (
-            _force_first_lookup_miss(OrganizationMembershipRequest.objects),
-            patch.object(OrganizationMembershipRequest, "full_clean", return_value=None),
-        ):
-            with pytest.raises(PendingMembershipRequestExistsError):
-                organization_service.create_membership_request(organization, nonmember_user)
-
-        surviving = OrganizationMembershipRequest.objects.filter(organization=organization, user=nonmember_user)
-        assert [row.pk for row in surviving] == [existing.pk]
-
-
-@pytest.mark.django_db
-class TestApproveMembershipRequest:
-    """Tests for the approve_membership_request function."""
-
-    def test_approve_membership_request_creates_member(
-        self, organization_membership_request: OrganizationMembershipRequest, organization_staff_user: RevelUser
-    ) -> None:
-        """Test that a member is created when a request is approved."""
-        # Arrange
-        from events.models import MembershipTier
-
-        tier = MembershipTier.objects.get(
-            organization=organization_membership_request.organization, name="General membership"
-        )
-
-        assert not OrganizationMember.objects.filter(
-            organization=organization_membership_request.organization, user=organization_membership_request.user
-        ).exists()
-
-        # Act
-        organization_service.approve_membership_request(organization_membership_request, organization_staff_user, tier)
-
-        # Assert
-        member = OrganizationMember.objects.get(
-            organization=organization_membership_request.organization, user=organization_membership_request.user
-        )
-        assert member is not None
-        assert member.tier == tier
-        assert member.status == OrganizationMember.MembershipStatus.ACTIVE
-        assert organization_membership_request.status == OrganizationMembershipRequest.Status.COMPLETED
-        assert organization_membership_request.decided_by == organization_staff_user
-
-    def test_approve_survives_lost_member_creation_race(
-        self, organization_membership_request: OrganizationMembershipRequest, organization_staff_user: RevelUser
-    ) -> None:
-        """A membership row committed concurrently must not 500 the approval.
-
-        ``(organization, user)`` is unique and ``TimeStampedModel.save`` runs
-        ``full_clean``, so a row created between the guards and our INSERT surfaces
-        as ``ValidationError`` — which Django's ``update_or_create`` does not absorb.
-        """
-        tier = MembershipTier.objects.get(
-            organization=organization_membership_request.organization, name="General membership"
-        )
-        existing = OrganizationMember.objects.create(
-            organization=organization_membership_request.organization,
-            user=organization_membership_request.user,
-            status=OrganizationMember.MembershipStatus.CANCELLED,
-        )
-
-        with patch.object(
-            OrganizationMember.objects,
-            "update_or_create",
-            side_effect=ValidationError("Organization member with this Organization and User already exists."),
-        ):
-            organization_service.approve_membership_request(
-                organization_membership_request, organization_staff_user, tier
-            )
-
-        existing.refresh_from_db()
-        assert existing.tier == tier
-        assert existing.status == OrganizationMember.MembershipStatus.ACTIVE
-        assert organization_membership_request.status == OrganizationMembershipRequest.Status.COMPLETED
-
-
-@pytest.mark.django_db
-class TestRejectMembershipRequest:
-    """Tests for the reject_membership_request function."""
-
-    def test_reject_membership_request_does_not_create_member(
-        self, organization_membership_request: OrganizationMembershipRequest, organization_staff_user: RevelUser
-    ) -> None:
-        """Test that a member is not created when a request is rejected."""
-        # Arrange
-        assert not OrganizationMember.objects.filter(
-            organization=organization_membership_request.organization, user=organization_membership_request.user
-        ).exists()
-
-        # Act
-        organization_service.reject_membership_request(organization_membership_request, organization_staff_user)
-
-        # Assert
-        assert not OrganizationMember.objects.filter(
-            organization=organization_membership_request.organization, user=organization_membership_request.user
-        ).exists()
-        assert organization_membership_request.status == OrganizationMembershipRequest.Status.REJECTED
-        assert organization_membership_request.decided_by == organization_staff_user
-
-
-@pytest.mark.django_db
-class TestClaimInvitation:
-    """Tests for the claim_invitation function."""
-
-    def test_claim_invitation_success(self, organization_token: OrganizationToken, nonmember_user: RevelUser) -> None:
-        """Test that an invitation is claimed successfully."""
-        # Act
-        claimed_org = organization_service.claim_invitation(nonmember_user, organization_token.id)
-
-        # Assert
-        assert claimed_org == organization_token.organization
-        assert OrganizationMember.objects.filter(
-            organization=organization_token.organization, user=nonmember_user
-        ).exists()
-        assert not OrganizationStaff.objects.filter(
-            organization=organization_token.organization, user=nonmember_user
-        ).exists()
-
-    def test_claim_invitation_staff_success(
-        self, staff_organization_token: OrganizationToken, nonmember_user: RevelUser
-    ) -> None:
-        """Test that a staff invitation is claimed successfully."""
-        # Act
-        claimed_org = organization_service.claim_invitation(nonmember_user, staff_organization_token.id)
-
-        # Assert
-        assert claimed_org == staff_organization_token.organization
-        assert OrganizationStaff.objects.filter(
-            organization=staff_organization_token.organization, user=nonmember_user
-        ).exists()
-        assert not OrganizationMember.objects.filter(
-            organization=staff_organization_token.organization, user=nonmember_user
-        ).exists()
-
-    def test_claim_invitation_grants_both_staff_and_membership(
-        self, organization: Organization, organization_owner_user: RevelUser, nonmember_user: RevelUser
-    ) -> None:
-        """A token granting both staff and membership must apply both, not just staff."""
-        default_tier = MembershipTier.objects.get(organization=organization, name="General membership")
-        token = OrganizationToken.objects.create(
-            organization=organization,
-            issuer=organization_owner_user,
-            grants_staff_status=True,
-            grants_membership=True,
-            membership_tier=default_tier,
-        )
-
-        # Act
-        claimed_org = organization_service.claim_invitation(nonmember_user, token.id)
-
-        # Assert - both grants applied, and the use is consumed exactly once
-        assert claimed_org == organization
-        assert OrganizationStaff.objects.filter(organization=organization, user=nonmember_user).exists()
-        member = OrganizationMember.objects.get(organization=organization, user=nonmember_user)
-        assert member.tier == default_tier
-        token.refresh_from_db()
-        assert token.uses == 1
-
-
-@pytest.mark.django_db
-class TestMemberManagement:
-    def test_add_member_success(self, organization: Organization, nonmember_user: RevelUser) -> None:
-        """Test that a user can be successfully added as a member."""
-        tier = MembershipTier.objects.create(organization=organization, name="Gold")
-        assert not OrganizationMember.objects.filter(organization=organization, user=nonmember_user).exists()
-        member = organization_service.add_member(organization, nonmember_user, tier)
-        assert member is not None
-        assert member.tier == tier
-        assert OrganizationMember.objects.filter(organization=organization, user=nonmember_user).exists()
-
-    def test_add_member_already_exists_fails(self, organization_membership: OrganizationMember) -> None:
-        """Test that adding an existing member raises an error."""
-        tier = MembershipTier.objects.create(organization=organization_membership.organization, name="Silver")
-        with pytest.raises(AlreadyMemberError):
-            organization_service.add_member(organization_membership.organization, organization_membership.user, tier)
-
-    def test_add_member_lost_race_raises_domain_error(self, organization_membership: OrganizationMember) -> None:
-        """A concurrent membership create must surface AlreadyMemberError, not a 500.
-
-        ``(organization, user)`` is unique and ``TimeStampedModel.save`` runs
-        ``full_clean``, so the bare ``create()`` used to blow up with
-        ``ValidationError`` when the pre-check ran too early to see the winner.
-        """
-        organization = organization_membership.organization
-        tier = MembershipTier.objects.create(organization=organization, name="Bronze")
-
-        with _force_first_lookup_miss(OrganizationMember.objects):
-            with pytest.raises(AlreadyMemberError):
-                organization_service.add_member(organization, organization_membership.user, tier)
-
-        surviving = OrganizationMember.objects.filter(organization=organization, user=organization_membership.user)
-        assert [row.pk for row in surviving] == [organization_membership.pk]
-
-    def test_add_member_lost_db_level_race_raises_domain_error(
-        self, organization_membership: OrganizationMember
-    ) -> None:
-        """The IntegrityError arm: ``full_clean`` disabled, so the unique index rejects the INSERT."""
-        organization = organization_membership.organization
-        tier = MembershipTier.objects.create(organization=organization, name="Copper")
-
-        with (
-            _force_first_lookup_miss(OrganizationMember.objects),
-            patch.object(OrganizationMember, "full_clean", return_value=None),
-        ):
-            with pytest.raises(AlreadyMemberError):
-                organization_service.add_member(organization, organization_membership.user, tier)
-
-        surviving = OrganizationMember.objects.filter(organization=organization, user=organization_membership.user)
-        assert [row.pk for row in surviving] == [organization_membership.pk]
-
-    def test_remove_member_success(self, organization_membership: OrganizationMember) -> None:
-        """Test that a member can be successfully removed."""
-        organization = organization_membership.organization
-        user = organization_membership.user
-        assert OrganizationMember.objects.filter(organization=organization, user=user).exists()
-        organization_service.remove_member(organization, user)
-        assert not OrganizationMember.objects.filter(organization=organization, user=user).exists()
-
-
-@pytest.mark.django_db
-class TestStaffManagement:
-    def test_add_staff_success(self, organization: Organization, nonmember_user: RevelUser) -> None:
-        """Test adding a staff member with default permissions."""
-        assert not OrganizationStaff.objects.filter(organization=organization, user=nonmember_user).exists()
-        staff = organization_service.add_staff(organization, nonmember_user)
-        assert staff is not None
-        assert staff.permissions is not None
-        assert OrganizationStaff.objects.filter(organization=organization, user=nonmember_user).exists()
-
-    def test_add_staff_with_custom_permissions(self, organization: Organization, nonmember_user: RevelUser) -> None:
-        """Test adding a staff member with custom permissions."""
-        custom_perms = PermissionsSchema(default=PermissionMap(create_event=True, edit_event=False))
-        staff = organization_service.add_staff(organization, nonmember_user, permissions=custom_perms)
-        assert staff.permissions["default"]["create_event"] is True
-        assert staff.permissions["default"]["edit_event"] is False
-
-    def test_add_staff_already_exists_fails(self, staff_member: OrganizationStaff) -> None:
-        """Test that adding an existing staff member raises an error."""
-        with pytest.raises(AlreadyMemberError):
-            organization_service.add_staff(staff_member.organization, staff_member.user)
-
-    def test_remove_staff_success(self, staff_member: OrganizationStaff) -> None:
-        """Test removing a staff member."""
-        organization = staff_member.organization
-        user = staff_member.user
-        assert OrganizationStaff.objects.filter(organization=organization, user=user).exists()
-        organization_service.remove_staff(organization, user)
-        assert not OrganizationStaff.objects.filter(organization=organization, user=user).exists()
-
-    def test_update_staff_permissions(self, staff_member: OrganizationStaff) -> None:
-        """Test updating a staff member's permissions."""
-        assert staff_member.has_permission("create_event") is False
-        new_perms = PermissionsSchema(default=PermissionMap(create_event=True))
-
-        updated_staff = organization_service.update_staff_permissions(staff_member, new_perms)
-        updated_staff.refresh_from_db()
-
-        assert updated_staff.has_permission("create_event") is True
-
-
-@pytest.mark.django_db(transaction=True)
-class TestCreateOrganization:
-    """Tests for the create_organization function."""
-
-    @patch("events.tasks.send_organization_contact_email_verification.delay")
-    def test_create_organization_success(self, mock_send_email: MagicMock, nonmember_user: RevelUser) -> None:
-        """Test that an organization is created successfully."""
-        # Arrange
-        nonmember_user.email_verified = True
-        nonmember_user.save()
-
-        # Act
-        organization = organization_service.create_organization(
-            owner=nonmember_user,
-            name="Acme Collective",
-            contact_email="contact@example.com",
-            description="Test description",
-        )
-
-        # Assert
-        assert organization.name == "Acme Collective"
-        assert organization.owner == nonmember_user
-        assert organization.description == "Test description"
-        assert organization.contact_email == "contact@example.com"
-        assert organization.contact_email_verified is False
-        assert organization.visibility == Organization.Visibility.STAFF_ONLY
-
-        # Check that verification email is sent
-        assert mock_send_email.called
-        call_args = mock_send_email.call_args[1]
-        assert call_args["email"] == "contact@example.com"
-        assert call_args["organization_name"] == "Acme Collective"
-        assert "token" in call_args
-
-    @patch("events.tasks.send_organization_contact_email_verification.delay")
-    def test_create_organization_with_owner_email_auto_verifies(
-        self, mock_send_email: MagicMock, nonmember_user: RevelUser
-    ) -> None:
-        """Test that contact email is auto-verified when it matches owner's verified email."""
-        # Arrange
-        nonmember_user.email_verified = True
-        nonmember_user.email = "owner@example.com"
-        nonmember_user.save()
-
-        # Act
-        organization = organization_service.create_organization(
-            owner=nonmember_user,
-            name="Acme Collective",
-            contact_email="owner@example.com",
-            description="Test description",
-        )
-
-        # Assert
-        assert organization.contact_email == "owner@example.com"
-        assert organization.contact_email_verified is True
-
-        # Check that no verification email is sent when auto-verified
-        assert not mock_send_email.called
-
-    def test_create_organization_user_already_owns_one_fails(self, organization: Organization) -> None:
-        """Test that a user cannot create a second organization."""
-        # Act & Assert
-        with pytest.raises(HttpError) as exc_info:
-            organization_service.create_organization(
-                owner=organization.owner,
-                name="Pebble Society",
-                contact_email="contact@example.com",
-            )
-        assert exc_info.value.status_code == 400
-        assert "already own an organization" in str(exc_info.value)
-
-    def test_create_organization_with_unverified_owner_email(self, nonmember_user: RevelUser) -> None:
-        """Test that contact email is not auto-verified when owner's email is unverified."""
-        # Arrange
-        nonmember_user.email_verified = False
-        nonmember_user.email = "owner@example.com"
-        nonmember_user.save()
-
-        # Act
-        organization = organization_service.create_organization(
-            owner=nonmember_user,
-            name="Acme Collective",
-            contact_email="owner@example.com",
-        )
-
-        # Assert
-        assert organization.contact_email == "owner@example.com"
-        assert organization.contact_email_verified is False
-
-    @patch("events.tasks.send_organization_contact_email_verification.delay")
-    def test_create_organization_rejects_hardcoded_reserved_token(
-        self, _mock_send_email: MagicMock, nonmember_user: RevelUser
-    ) -> None:
-        """Test that the service rejects names containing a hardcoded reserved token."""
-        with pytest.raises(HttpError) as exc_info:
-            organization_service.create_organization(
-                owner=nonmember_user,
-                name="Test Organization",
-                contact_email="contact@example.com",
-            )
-        assert exc_info.value.status_code == 400
-        assert "test" in str(exc_info.value).lower()
-
-    @patch("events.tasks.send_organization_contact_email_verification.delay")
-    def test_create_organization_rejects_word_order_variant(
-        self, _mock_send_email: MagicMock, nonmember_user: RevelUser
-    ) -> None:
-        """Test that the guard catches the reserved token regardless of position."""
-        with pytest.raises(HttpError) as exc_info:
-            organization_service.create_organization(
-                owner=nonmember_user,
-                name="Choir Test",
-                contact_email="contact@example.com",
-            )
-        assert exc_info.value.status_code == 400
-
-    @patch("events.tasks.send_organization_contact_email_verification.delay")
-    def test_create_organization_rejects_db_token(self, _mock_send_email: MagicMock, nonmember_user: RevelUser) -> None:
-        """Test that the guard reads from the DB-backed reserved-token list."""
-        from events.models import ReservedSlugToken
-        from events.utils.reserved_slug_tokens import invalidate_reserved_tokens_cache
-
-        ReservedSlugToken.objects.create(token="forbidden", reason="")
-        invalidate_reserved_tokens_cache()
-        with pytest.raises(HttpError):
-            organization_service.create_organization(
-                owner=nonmember_user,
-                name="My Forbidden Club",
-                contact_email="contact@example.com",
-            )
-
-    @patch("events.tasks.send_organization_contact_email_verification.delay")
-    def test_create_organization_allows_clean_name(
-        self, _mock_send_email: MagicMock, nonmember_user: RevelUser
-    ) -> None:
-        """Test that a name with no reserved tokens passes the guard."""
-        org = organization_service.create_organization(
-            owner=nonmember_user,
-            name="Acoustic Events Collective",
-            contact_email="contact@example.com",
-        )
-        assert org.pk is not None
-
-
-@pytest.mark.django_db(transaction=True)
-class TestUpdateContactEmail:
-    """Tests for the update_contact_email function."""
-
-    @patch("events.tasks.send_organization_contact_email_verification.delay")
-    def test_update_contact_email_success(
-        self, mock_send_email: MagicMock, organization: Organization, organization_owner_user: RevelUser
-    ) -> None:
-        """Test updating contact email successfully."""
-        # Act
-        token = organization_service.update_contact_email(
-            organization=organization,
-            new_email="newemail@example.com",
-            requester=organization_owner_user,
-        )
-
-        # Assert
-        organization.refresh_from_db()
-        assert organization.contact_email == "newemail@example.com"
-        assert organization.contact_email_verified is False
-        assert token != ""
-        assert mock_send_email.called
-        mock_send_email.assert_called_once_with(
-            email="newemail@example.com",
-            token=token,
-            organization_name=organization.name,
-            organization_slug=organization.slug,
-        )
-
-    @patch("events.tasks.send_organization_contact_email_verification.delay")
-    def test_update_contact_email_auto_verifies_with_user_email(
-        self, mock_send_email: MagicMock, organization: Organization, organization_owner_user: RevelUser
-    ) -> None:
-        """Test that contact email is auto-verified when it matches requester's verified email."""
-        # Arrange
-        organization_owner_user.email_verified = True
-        organization_owner_user.email = "owner@example.com"
-        organization_owner_user.save()
-
-        # Act
-        token = organization_service.update_contact_email(
-            organization=organization,
-            new_email="owner@example.com",
-            requester=organization_owner_user,
-        )
-
-        # Assert
-        organization.refresh_from_db()
-        assert organization.contact_email == "owner@example.com"
-        assert organization.contact_email_verified is True
-        assert token == ""  # No token needed when auto-verified
-
-        # Check that no email is sent when auto-verified
-        assert not mock_send_email.called
-
-    def test_update_contact_email_same_email_fails(
-        self, organization: Organization, organization_owner_user: RevelUser
-    ) -> None:
-        """Test that updating to the same email fails."""
-        # Arrange
-        organization.contact_email = "existing@example.com"
-        organization.save()
-
-        # Act & Assert
-        with pytest.raises(HttpError) as exc_info:
-            organization_service.update_contact_email(
-                organization=organization,
-                new_email="existing@example.com",
-                requester=organization_owner_user,
-            )
-        assert exc_info.value.status_code == 400
-        assert "already the contact email" in str(exc_info.value)
-
-
-@pytest.mark.django_db
-class TestVerifyContactEmail:
-    """Tests for the verify_contact_email function."""
-
-    def test_verify_contact_email_success(self, organization: Organization, organization_owner_user: RevelUser) -> None:
-        """Test verifying contact email with valid token."""
-        # Arrange
-        organization.contact_email = "test@example.com"
-        organization.contact_email_verified = False
-        organization.save()
-
-        # Create a valid token
-        verification_payload = schema.VerifyOrganizationContactEmailJWTPayloadSchema(
-            organization_id=organization.id,
-            user_id=organization_owner_user.id,
-            email="test@example.com",
-            exp=timezone.now() + settings.VERIFY_TOKEN_LIFETIME,
-        )
-        token = create_token(verification_payload.model_dump(mode="json"), settings.SECRET_KEY, settings.JWT_ALGORITHM)
-
-        # Act
-        verified_org = organization_service.verify_contact_email(token)
-
-        # Assert
-        assert verified_org.contact_email_verified is True
-        assert verified_org.id == organization.id
-
-    def test_verify_contact_email_wrong_email_fails(
-        self, organization: Organization, organization_owner_user: RevelUser
-    ) -> None:
-        """Test that verification fails when email has changed."""
-        # Arrange
-        organization.contact_email = "current@example.com"
-        organization.contact_email_verified = False
-        organization.save()
-
-        # Create a token for a different email
-        verification_payload = schema.VerifyOrganizationContactEmailJWTPayloadSchema(
-            organization_id=organization.id,
-            user_id=organization_owner_user.id,
-            email="old@example.com",
-            exp=timezone.now() + settings.VERIFY_TOKEN_LIFETIME,
-        )
-        token = create_token(verification_payload.model_dump(mode="json"), settings.SECRET_KEY, settings.JWT_ALGORITHM)
-
-        # Act & Assert
-        with pytest.raises(HttpError) as exc_info:
-            organization_service.verify_contact_email(token)
-        assert exc_info.value.status_code == 400
-        assert "different email address" in str(exc_info.value)
-
-    def test_verify_contact_email_invalid_organization_fails(self, organization_owner_user: RevelUser) -> None:
-        """Test that verification fails for non-existent organization."""
-        # Arrange - Create a token for non-existent organization
-        from uuid import uuid4
-
-        verification_payload = schema.VerifyOrganizationContactEmailJWTPayloadSchema(
-            organization_id=uuid4(),  # Non-existent ID
-            user_id=organization_owner_user.id,
-            email="test@example.com",
-            exp=timezone.now() + settings.VERIFY_TOKEN_LIFETIME,
-        )
-        token = create_token(verification_payload.model_dump(mode="json"), settings.SECRET_KEY, settings.JWT_ALGORITHM)
-
-        # Act & Assert
-        with pytest.raises(HttpError) as exc_info:
-            organization_service.verify_contact_email(token)
-        assert exc_info.value.status_code == 400
-        assert "Organization not found" in str(exc_info.value)
-
-    def test_verify_contact_email_blacklisted_token_fails(
-        self, organization: Organization, organization_owner_user: RevelUser
-    ) -> None:
-        """Test that verification fails with blacklisted token."""
-        # Arrange
-        organization.contact_email = "test@example.com"
-        organization.contact_email_verified = False
-        organization.save()
-
-        # Create and blacklist a token
-        verification_payload = schema.VerifyOrganizationContactEmailJWTPayloadSchema(
-            organization_id=organization.id,
-            user_id=organization_owner_user.id,
-            email="test@example.com",
-            exp=timezone.now() + settings.VERIFY_TOKEN_LIFETIME,
-        )
-        token = create_token(verification_payload.model_dump(mode="json"), settings.SECRET_KEY, settings.JWT_ALGORITHM)
-        blacklist_token(token)
-
-        # Act & Assert
-        with pytest.raises(HttpError):
-            organization_service.verify_contact_email(token)
-
-
-@pytest.mark.django_db
-class TestCreateOrganizationTokenValidation:
-    """Tests for M-02: organization tokens must grant at least one type of access."""
-
-    def test_create_token_with_grants_membership_succeeds(
-        self, organization: Organization, organization_owner_user: RevelUser
-    ) -> None:
-        """Token with grants_membership=True can be created via service."""
-        default_tier = MembershipTier.objects.get(organization=organization, name="General membership")
-        token = organization_service.create_organization_token(
-            organization=organization,
-            issuer=organization_owner_user,
-            grants_membership=True,
-            grants_staff_status=False,
-            membership_tier=default_tier,
-        )
-        assert token.grants_membership is True
-        assert token.grants_staff_status is False
-
-    def test_create_token_with_grants_staff_status_succeeds(
-        self, organization: Organization, organization_owner_user: RevelUser
-    ) -> None:
-        """Token with grants_staff_status=True can be created via service."""
-        token = organization_service.create_organization_token(
-            organization=organization,
-            issuer=organization_owner_user,
-            grants_membership=False,
-            grants_staff_status=True,
-        )
-        assert token.grants_membership is False
-        assert token.grants_staff_status is True
-
-    def test_create_token_with_both_grants_succeeds(
-        self, organization: Organization, organization_owner_user: RevelUser
-    ) -> None:
-        """Token with both grants enabled can be created via service."""
-        default_tier = MembershipTier.objects.get(organization=organization, name="General membership")
-        token = organization_service.create_organization_token(
-            organization=organization,
-            issuer=organization_owner_user,
-            grants_membership=True,
-            grants_staff_status=True,
-            membership_tier=default_tier,
-        )
-        assert token.grants_membership is True
-        assert token.grants_staff_status is True
-
-    def test_create_token_with_no_grants_raises_value_error(
-        self, organization: Organization, organization_owner_user: RevelUser
-    ) -> None:
-        """Token with both grants disabled raises ValueError in service."""
-        with pytest.raises(ValueError, match="At least one of grants_membership or grants_staff_status must be True"):
-            organization_service.create_organization_token(
-                organization=organization,
-                issuer=organization_owner_user,
-                grants_membership=False,
-                grants_staff_status=False,
-            )
-
-
-@pytest.mark.django_db
-class TestCreateOrganizationTokenFromPayload:
-    """Tests for ``organization_service.create_organization_token_from_payload``."""
-
-    def test_owner_can_create_staff_granting_token(
-        self, organization: Organization, organization_owner_user: RevelUser
-    ) -> None:
-        default_tier = MembershipTier.objects.get(organization=organization, name="General membership")
-        payload = schema.OrganizationTokenCreateSchema(
-            name="Staff Token",
-            grants_membership=True,
-            grants_staff_status=True,
-            membership_tier_id=default_tier.id,
-        )
-
-        token = organization_service.create_organization_token_from_payload(
-            organization=organization, requested_by=organization_owner_user, payload=payload
-        )
-
-        assert token.grants_staff_status is True
-        assert token.membership_tier_id == default_tier.id
-
-    def test_non_owner_cannot_create_staff_granting_token(
-        self, organization: Organization, organization_staff_user: RevelUser
-    ) -> None:
-        default_tier = MembershipTier.objects.get(organization=organization, name="General membership")
-        payload = schema.OrganizationTokenCreateSchema(
-            name="Staff Token",
-            grants_membership=True,
-            grants_staff_status=True,
-            membership_tier_id=default_tier.id,
-        )
-
-        with pytest.raises(OrganizationTokenStaffGrantForbidden):
-            organization_service.create_organization_token_from_payload(
-                organization=organization, requested_by=organization_staff_user, payload=payload
-            )
-
-    def test_non_owner_can_create_membership_only_token(
-        self, organization: Organization, organization_staff_user: RevelUser
-    ) -> None:
-        default_tier = MembershipTier.objects.get(organization=organization, name="General membership")
-        payload = schema.OrganizationTokenCreateSchema(
-            name="Member Token",
-            grants_membership=True,
-            grants_staff_status=False,
-            membership_tier_id=default_tier.id,
-        )
-
-        token = organization_service.create_organization_token_from_payload(
-            organization=organization, requested_by=organization_staff_user, payload=payload
-        )
-
-        assert token.grants_membership is True
-        assert token.grants_staff_status is False
-
-
-@pytest.mark.django_db
-class TestUpdateOrganizationTokenService:
-    """Tests for the high-level ``organization_service.update_organization_token``."""
-
-    def test_grant_invariant_violation_raises(
-        self, organization: Organization, staff_organization_token: OrganizationToken
-    ) -> None:
-        # The staff token already has grants_membership=False. A partial update
-        # that flips grants_staff_status=False would leave both False, which the
-        # schema's "both explicitly set" validator misses (only one field is in
-        # model_fields_set). The service-side check catches it as defense-in-depth.
-        payload = schema.OrganizationTokenUpdateSchema(grants_staff_status=False)
-
-        with pytest.raises(OrganizationTokenGrantInvariantError):
-            organization_service.update_organization_token(
-                staff_organization_token, requested_by=organization.owner, payload=payload
-            )
-
-        staff_organization_token.refresh_from_db()
-        assert staff_organization_token.grants_staff_status is True
-
-    def test_non_owner_cannot_promote_token_to_staff(
-        self,
-        organization: Organization,
-        organization_token: OrganizationToken,
-        organization_staff_user: RevelUser,
-    ) -> None:
-        default_tier = MembershipTier.objects.get(organization=organization, name="General membership")
-        payload = schema.OrganizationTokenUpdateSchema(
-            grants_staff_status=True, grants_membership=True, membership_tier_id=default_tier.id
-        )
-
-        with pytest.raises(OrganizationTokenStaffGrantForbidden):
-            organization_service.update_organization_token(
-                organization_token, requested_by=organization_staff_user, payload=payload
-            )
-
-        organization_token.refresh_from_db()
-        assert organization_token.grants_staff_status is False
-
-    def test_clearing_membership_tier_while_grants_membership_raises(
-        self, organization: Organization, organization_token: OrganizationToken
-    ) -> None:
-        # The token has grants_membership=True with a tier. A partial update
-        # that only sets membership_tier_id=None slips past the schema validator
-        # (grants_membership is not in model_fields_set), but would leave the
-        # token in an inconsistent state that OrganizationToken.clean() rejects
-        # at full_clean() time — surfacing as a 500. The service-side check
-        # raises a structured exception the controller maps to 422.
-        payload = schema.OrganizationTokenUpdateSchema(membership_tier_id=None)
-
-        with pytest.raises(OrganizationTokenMembershipTierRequiredError):
-            organization_service.update_organization_token(
-                organization_token, requested_by=organization.owner, payload=payload
-            )
-
-        organization_token.refresh_from_db()
-        assert organization_token.membership_tier_id is not None
-
-    def test_non_owner_cannot_touch_existing_staff_token(
-        self,
-        organization: Organization,
-        staff_organization_token: OrganizationToken,
-        organization_staff_user: RevelUser,
-    ) -> None:
-        payload = schema.OrganizationTokenUpdateSchema(name="renamed")
-
-        with pytest.raises(OrganizationTokenStaffGrantForbidden):
-            organization_service.update_organization_token(
-                staff_organization_token, requested_by=organization_staff_user, payload=payload
-            )
-
-    def test_owner_can_update_membership_tier(
-        self,
-        organization: Organization,
-        organization_token: OrganizationToken,
-        organization_owner_user: RevelUser,
-    ) -> None:
-        new_tier = MembershipTier.objects.create(organization=organization, name="VIP")
-        payload = schema.OrganizationTokenUpdateSchema(membership_tier_id=new_tier.id)
-
-        updated = organization_service.update_organization_token(
-            organization_token, requested_by=organization_owner_user, payload=payload
-        )
-
-        assert updated.membership_tier_id == new_tier.id
-
-
-@pytest.mark.django_db
-class TestDeleteOrganizationTokenService:
-    """Tests for the high-level ``organization_service.delete_organization_token``."""
-
-    def test_owner_can_delete_staff_token(
-        self,
-        organization_owner_user: RevelUser,
-        staff_organization_token: OrganizationToken,
-    ) -> None:
-        token_id = staff_organization_token.id
-        organization_service.delete_organization_token(staff_organization_token, requested_by=organization_owner_user)
-        assert not OrganizationToken.objects.filter(pk=token_id).exists()
-
-    def test_non_owner_cannot_delete_staff_token(
-        self,
-        staff_organization_token: OrganizationToken,
-        organization_staff_user: RevelUser,
-    ) -> None:
-        with pytest.raises(OrganizationTokenStaffGrantForbidden):
-            organization_service.delete_organization_token(
-                staff_organization_token, requested_by=organization_staff_user
-            )
-        assert OrganizationToken.objects.filter(pk=staff_organization_token.id).exists()
-
-    def test_non_owner_can_delete_membership_only_token(
-        self,
-        organization_token: OrganizationToken,
-        organization_staff_user: RevelUser,
-    ) -> None:
-        token_id = organization_token.id
-        organization_service.delete_organization_token(organization_token, requested_by=organization_staff_user)
-        assert not OrganizationToken.objects.filter(pk=token_id).exists()
